Develop

+++ neurolabi/gui/command/zstackdownsamplecommand.cpp
@@ -6,11 +6,20 @@
 #include "zstack.hxx"
 #include "zstackwriter.h"
 
+
 ZStackDownsampleCommand::ZStackDownsampleCommand()
 {
 
 }
 
+/*!
+ * Configuration
+ * {
+ *   "command": "downsample_stack",
+ *   "intv": [<int>, <int>, <int>],
+ *   "option": "max"|"min"|"mean"
+ * }
+ */
 int ZStackDownsampleCommand::run(
     const std::vector<std::string> &input, const std::string &output,
     const ZJsonObject &config)
@@ -19,7 +28,7 @@ int ZStackDownsampleCommand::run(
 
   if (!input.empty() && !output.empty()) {
     if (config.hasKey("intv")) {
-      std::vector<int> intv = ZJsonParser::integerArray(config["intv"]);
+      std::vector<int64_t> intv = ZJsonParser::integerArray(config["intv"]);
       if (intv.size() == 3) {
         if (intv[0] >= 0 && intv[1] >= 0 && intv[2] >= 0) {
           ZStack stack;
